run capsule, asset role viewable, rename file from circular import
