+++ src/npc_lims/paths/codeocean_s3.py
